CFY-7123 Handle the case of no workdir (#639)
* get_profile_dir: if suppress_error, actually return None on errors

Otherwise, even with suppress_error set, get_profile_dir throws
a TypeError if active_profile is None

* get_context_path: handle the case of no profile dir

Pass suppress_error through: if we get none from get_profile_dir,
that means suppress_error is set, so we can just return None

Callers of get_context_path need to handle that case, of course:
there's only one caller setting suppress_error: init_manager_profile

* init_manager_profile: handle the case of no context_file_path

We set suppress_error=True, so we need to handle the case of no
context_file_path

* get_default_rest_cert_local_path: handle profile dir=None

* get_profile_dir with suppress_error unittests

* More unittests

+++ cloudify_cli/commands/init.py
@@ -155,7 +155,7 @@ def init_manager_profile(profile_name,
 
     context_file_path = env.get_context_path(profile_name, suppress_error=True)
 
-    if os.path.isfile(context_file_path):
+    if context_file_path and os.path.isfile(context_file_path):
         if reset_context:
             if hard:
                 os.remove(config.CLOUDIFY_CONFIG_PATH)
